requirements.txt
Flask
python-dotenv
watchdog 